+++ DevExpress/QLBV_DEV/Thuoc.cs
@@ -26,7 +26,7 @@ namespace QLBV_DEV
         public Nullable<double> GiaBanLe { get; set; }
         public Nullable<double> GiaBanBuon { get; set; }
         public Nullable<double> QuyCach { get; set; }
-        public Nullable<int> TonKho { get; set; }
+        public Nullable<double> TonKho { get; set; }
         public Nullable<int> ThoiGianCanhBaoHetHan { get; set; }
         public Nullable<int> TonKhoToiThieu { get; set; }
         public Nullable<System.DateTime> NgayTao { get; set; }
